requirements.txt
flask_login==0.5.0
flask_migrate==2.7.0
flask_wtf==0.14.3
sqlalchemy==1.3.23
email_validator==1.1.2
alembic==1.7.4
black==21.9b0
cachelib==0.4.1
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
PyMySQL==1.0.2
python-decouple==3.4
python-dotenv==0.19.0
marshmallow==3.14.0
Jinja2==3.0.2
phonenumbers==8.11.0
regex==2021.10.8
SQLAlchemy==1.3.23
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
flask-cors==3.0.10
requests==2.21.0
pyOpenSSL==19.0.0
oauthlib==3.0.1